Implement aggregation for counters
Slowly getting the hang of it, I need to read up on the different
traits, such as Clone and Copy. Initially I didn't have Copy on the
MessageKind enum, which lead to some problems were values were moved. I
think enums are represented quite small in rust, so Copy should be fine.

For the hashmap I probably don't want to store the actual metric names
as keys, because I have those in the Metric struct already.