Flush metrics periodically to carbon
I knew from the start that this part would be 'challenging' because of
rust's borrow checker. For now, I opted for the easy approach, by just
wrapping the whole aggregator in an arc/mutex. Initially I wanted to
handle the thread inside the aggregator, but all I got was lifetime
issues. I don't blame rust here, because it is correct, but that still
was a hurdle.

Will revisit this later.

Now I have a working statsd that aggregates counters and flushes to
carbon periodically, yay!